refactoring due to review
