+++ kafka-ui-e2e-checks/src/main/java/com/provectus/kafka/ui/screenshots/NoReferenceScreenshotFoundException.java
@@ -1,0 +1,7 @@
+package com.provectus.kafka.ui.screenshots;
+
+public class NoReferenceScreenshotFoundException extends Throwable {
+    public NoReferenceScreenshotFoundException(String name) {
+        super("no reference screenshot found for " + name);
+    }
+}
