+++ kafka-ui-e2e-checks/src/main/java/com/provectus/kafka/ui/utils/qaseIO/annotation/Suite.java
@@ -1,0 +1,10 @@
+package com.provectus.kafka.ui.utils.qaseIO.annotation;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Suite {
+    long suiteId();
+    String title();
+}
