--- kafka-ui-e2e-checks/src/main/java/screenshots/NoReferenceScreenshotFoundException.java
@@ -1,7 +1,0 @@
-package screenshots;
-
-public class NoReferenceScreenshotFoundException extends Throwable {
-    public NoReferenceScreenshotFoundException(String name) {
-        super("no reference screenshot found for " + name);
-    }
-}
