--- kafka-ui-e2e-checks/src/main/java/utils/qaseIO/annotation/Suite.java
@@ -1,10 +1,0 @@
-package utils.qaseIO.annotation;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Suite {
-    long suiteId() default 0;
-    String title() default "";
-}
