+++ kafka-ui-e2e-checks/src/main/resources/META-INF/services/org.junit.platform.launcher.TestExecutionListener
@@ -1,1 +1,1 @@
-extensions.QaseExtension
+com.provectus.kafka.ui.extensions.QaseExtension
