+++ kafka-ui-e2e-checks/src/test/java/com/provectus/kafka/ui/base/BaseTest.java
@@ -3,8 +3,8 @@ package com.provectus.kafka.ui.base;
 import com.codeborne.selenide.Configuration;
 import com.codeborne.selenide.WebDriverRunner;
 import com.codeborne.selenide.logevents.SelenideLogger;
-import helpers.Helpers;
-import helpers.TestConfiguration;
+import com.provectus.kafka.ui.helpers.Helpers;
+import com.provectus.kafka.ui.helpers.TestConfiguration;
 import io.github.cdimascio.dotenv.Dotenv;
 import io.qameta.allure.Allure;
 import io.qameta.allure.selenide.AllureSelenide;
@@ -21,9 +21,9 @@ import org.testcontainers.Testcontainers;
 import org.testcontainers.containers.BrowserWebDriverContainer;
 import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.utility.DockerImageName;
-import pages.Pages;
-import screenshots.Screenshooter;
-import utils.CamelCaseToSpacedDisplayNameGenerator;
+import com.provectus.kafka.ui.pages.Pages;
+import com.provectus.kafka.ui.screenshots.Screenshooter;
+import com.provectus.kafka.ui.utils.CamelCaseToSpacedDisplayNameGenerator;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
